Add action performed on colors

+++ Frontend/src/Components/ColorMenu.tsx
@@ -4,7 +4,7 @@ import { ThemeContext, useColorMenuContext } from "../context";
 import { IconPalette, IconPlus, IconTrash } from "@tabler/icons-react";
 import "./ColorMenuStyles.css"
 
-export default function ColorMenu() {
+export default function ColorMenu(props: { changeSelectionsToDefaultColorByColorID: (colorID: number) => void }) {
     const [colorMenu, setColorMenu] = useState(false);
     const [newColorIDCounter, setNewColorIDCounter] = useState(-1);
     const [currentDescription, setCurrentDescription] = useState("");
@@ -16,6 +16,7 @@ export default function ColorMenu() {
         const newSavedColors = colorMenuContext.savedColors.map((color) => {
             if (color.Color_id === option.Color_id) {
                 color.Hex_value = value;
+                color.Action = "changed";
             }
             return color;
         });
@@ -23,7 +24,7 @@ export default function ColorMenu() {
     }
 
     const addColor = () => {
-        colorMenuContext.setSavedColors([...colorMenuContext.savedColors, { Color_id: newColorIDCounter, Description: "Default", Hex_value: "#747bff" }]);
+        colorMenuContext.setSavedColors([...colorMenuContext.savedColors, { Action: "created", Color_id: newColorIDCounter, Description: "New Color", Hex_value: "#747bff" }]);
         colorMenuContext.setSelectedColorID(newColorIDCounter);
         setNewColorIDCounter(newColorIDCounter - 1);
     }
@@ -41,7 +42,20 @@ export default function ColorMenu() {
             }
         }
 
-        colorMenuContext.setSavedColors(colorMenuContext.savedColors.filter((color) => color.Color_id !== colorMenuContext.selectedColorID));
+        if (toDelete.Action === "saved") {
+            const newSavedColors = colorMenuContext.savedColors.map((color) => {
+                if (color.Color_id === toDelete.Color_id) {
+                    color.Action = "deleted";
+                }
+                return color;
+            });
+            colorMenuContext.setSavedColors(newSavedColors);
+        }
+        else {
+            colorMenuContext.setSavedColors(colorMenuContext.savedColors.filter((color) => color.Color_id !== colorMenuContext.selectedColorID));
+        }
+
+        props.changeSelectionsToDefaultColorByColorID(toDelete.Color_id);
     }
 
     const onDescriptionChange = (value: string) => {
@@ -52,6 +66,7 @@ export default function ColorMenu() {
         const newSavedColors = colorMenuContext.savedColors.map((color) => {
             if (color.Color_id === colorMenuContext.selectedColorID) {
                 color.Description = currentDescription;
+                color.Action = color.Action === "created" ? "created" : "updated";
             }
             return color;
         });
@@ -86,36 +101,40 @@ export default function ColorMenu() {
                         <IconPlus onClick={() => addColor()} />
                     </div>
 
-                    {colorMenuContext.savedColors.map((colorOption) =>
-                        <div
-                            className={"colorMenuOptions " + ((colorOption.Color_id === colorMenuContext.selectedColorID) ? "selectedColorOption" : "")}
-                            key={colorOption.Color_id}
-                            onClick={() => colorMenuContext.setSelectedColorID(colorOption.Color_id)}
-                        >
-                            {colorOption.Color_id === colorMenuContext.selectedColorID &&
+                    {colorMenuContext.savedColors.map((colorOption) => {
+                        if (colorOption.Action !== "deleted") {
+                            return <div
+                                className={"colorMenuOptions " + ((colorOption.Color_id === colorMenuContext.selectedColorID) ? "selectedColorOption" : "")}
+                                key={colorOption.Color_id}
+                                onClick={() => colorMenuContext.setSelectedColorID(colorOption.Color_id)}
+                            >
+                                {colorOption.Color_id === colorMenuContext.selectedColorID &&
+                                    <input
+                                        type="text"
+                                        className="description"
+                                        defaultValue={colorOption.Description}
+                                        onFocus={() => setCurrentDescription(colorOption.Description)}
+                                        onChange={(e) => onDescriptionChange(e.target.value)}
+                                        onBlur={() => onDescriptionBlur()}
+                                        disabled={colorOption.Action === "default"}
+                                    />
+                                }
+                                {colorOption.Color_id !== colorMenuContext.selectedColorID &&
+                                    <div className="description">{colorOption.Description}</div>
+                                }
                                 <input
-                                    type="text"
-                                    className="description"
-                                    defaultValue={colorOption.Description}
-                                    onFocus={() => setCurrentDescription(colorOption.Description)}
-                                    onChange={(e) => onDescriptionChange(e.target.value)}
-                                    onBlur={() => onDescriptionBlur()}
+                                    className="colorPicker"
+                                    type="color"
+                                    value={colorOption.Hex_value}
+                                    onChange={e => onColorChange(colorOption, e.target.value)}
+                                    disabled={colorOption.Action === "default"}
                                 />
-                            }
-                            {colorOption.Color_id !== colorMenuContext.selectedColorID &&
-                                <div className="description">{colorOption.Description}</div>
-                            }
-                            <input
-                                className="colorPicker"
-                                type="color"
-                                value={colorOption.Hex_value}
-                                onChange={e => onColorChange(colorOption, e.target.value)}
-                            />
-                            {colorOption.Color_id === colorMenuContext.selectedColorID &&
-                                <IconTrash onClick={(e) => deleteColor(e, colorOption)} />
-                            }
-                        </div>
-                    )}
+                                {colorOption.Color_id === colorMenuContext.selectedColorID && colorOption.Action !== "default" &&
+                                    <IconTrash onClick={(e) => deleteColor(e, colorOption)} />
+                                }
+                            </div>
+                        }
+                    })}
                 </div>
             }
         </div>
